ipam: Decouple ENI device configuration from CRD allocator
Extract ENI device configuration into a standalone CiliumNode observer
that runs independently of the IPAM allocator. Previously,
`configureENIDevices` was called from
`nodeStore.updateLocalNodeResource`, coupling ENI network device setup to
the CRD allocator code path.

The new observer follows the same `job.Observer` pattern used by
`startLocalNodeAllocCIDRsSync` in the multi-pool allocator. It watches
CiliumNode updates and configures newly attached ENI devices regardless
of which allocator is active.

Relates to [cilium/design-cfps#87](https://github.com/cilium/design-cfps/pull/87)
