+++ pkg/ipam/crd_test.go
@@ -4,24 +4,28 @@
 package ipam
 
 import (
+	"context"
 	"errors"
 	"fmt"
+	"log/slog"
 	"net"
 	"net/netip"
 	"testing"
 	"time"
 
+	"github.com/cilium/hive/cell"
 	"github.com/cilium/hive/hivetest"
+	"github.com/cilium/hive/job"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	eniTypes "github.com/cilium/cilium/pkg/aws/eni/types"
 	azureTypes "github.com/cilium/cilium/pkg/azure/types"
 	fakeTypes "github.com/cilium/cilium/pkg/datapath/fake/types"
+	"github.com/cilium/cilium/pkg/hive"
 	ipamOption "github.com/cilium/cilium/pkg/ipam/option"
 	ipamTypes "github.com/cilium/cilium/pkg/ipam/types"
 	"github.com/cilium/cilium/pkg/ipmasq"
-	ciliumv2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
 	"github.com/cilium/cilium/pkg/logging"
 	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/node"
@@ -182,6 +186,14 @@ func TestIPMasq(t *testing.T) {
 	sharedNodeStore = newFakeNodeStore(conf, t)
 	sharedNodeStore.ownNode = cn
 
+	var jg job.Group
+	h := hive.New(
+		cell.Invoke(func(jg_ job.Group) { jg = jg_ }),
+	)
+	tlog := hivetest.Logger(t, hivetest.LogLevel(slog.LevelError))
+	require.NoError(t, h.Start(tlog, t.Context()))
+	t.Cleanup(func() { h.Stop(tlog, context.Background()) })
+
 	localNodeStore := node.NewTestLocalNodeStore(node.LocalNode{})
 	ipam := NewIPAM(NewIPAMParams{
 		Logger:         hivetest.Logger(t),
@@ -193,6 +205,7 @@ func TestIPMasq(t *testing.T) {
 		NodeResource:   &resourceMock{},
 		MTUConfig:      &mtuMock,
 		IPMasqAgent:    ipMasqAgent,
+		JobGroup:       jg,
 	})
 	ipam.ConfigureAllocator()
 
@@ -296,213 +309,4 @@ func TestAzureIPMasq(t *testing.T) {
 	)
 
 	ipMasqAgent.Stop()
-}
-
-func Test_validateENIConfig(t *testing.T) {
-	type args struct {
-		node *ciliumv2.CiliumNode
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-		want    string
-	}{
-		{
-			name: "Consistent ENI config",
-			args: args{
-				node: &ciliumv2.CiliumNode{
-					Spec: ciliumv2.NodeSpec{
-						IPAM: ipamTypes.IPAMSpec{
-							Pool: ipamTypes.AllocationMap{
-								"10.1.1.226": ipamTypes.AllocationIP{
-									Resource: "eni-1",
-								},
-							},
-						},
-					},
-					Status: ciliumv2.NodeStatus{
-						ENI: eniTypes.ENIStatus{
-							ENIs: map[string]eniTypes.ENI{
-								"eni-1": {
-									ID: "eni-1",
-									Addresses: []string{
-										"10.1.1.226",
-										"10.1.1.229",
-									},
-									VPC: eniTypes.AwsVPC{
-										ID:          "vpc-1",
-										PrimaryCIDR: "10.1.0.0/16",
-										CIDRs: []string{
-											"10.1.0.0/16",
-											"10.2.0.0/16",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "Missing VPC Primary CIDR",
-			args: args{
-				node: &ciliumv2.CiliumNode{
-					Spec: ciliumv2.NodeSpec{
-						IPAM: ipamTypes.IPAMSpec{
-							Pool: ipamTypes.AllocationMap{
-								"10.1.1.226": ipamTypes.AllocationIP{
-									Resource: "eni-1",
-								},
-							},
-						},
-					},
-					Status: ciliumv2.NodeStatus{
-						ENI: eniTypes.ENIStatus{
-							ENIs: map[string]eniTypes.ENI{
-								"eni-1": {
-									ID: "eni-1",
-									Addresses: []string{
-										"10.1.1.226",
-										"10.1.1.229",
-									},
-									VPC: eniTypes.AwsVPC{
-										ID: "vpc-1",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			wantErr: true,
-			want:    "VPC Primary CIDR not set for ENI eni-1",
-		},
-		{
-			name: "VPC CIDRs contain invalid value",
-			args: args{
-				node: &ciliumv2.CiliumNode{
-					Spec: ciliumv2.NodeSpec{
-						IPAM: ipamTypes.IPAMSpec{
-							Pool: ipamTypes.AllocationMap{
-								"10.1.1.226": ipamTypes.AllocationIP{
-									Resource: "eni-1",
-								},
-							},
-						},
-					},
-					Status: ciliumv2.NodeStatus{
-						ENI: eniTypes.ENIStatus{
-							ENIs: map[string]eniTypes.ENI{
-								"eni-1": {
-									ID: "eni-1",
-									Addresses: []string{
-										"10.1.1.226",
-										"10.1.1.229",
-									},
-									VPC: eniTypes.AwsVPC{
-										ID:          "vpc-1",
-										PrimaryCIDR: "10.1.0.0/16",
-										CIDRs: []string{
-											"",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			wantErr: true,
-			want:    "VPC CIDR not set for ENI eni-1",
-		},
-		{
-			name: "ENI not found in status",
-			args: args{
-				node: &ciliumv2.CiliumNode{
-					Spec: ciliumv2.NodeSpec{
-						IPAM: ipamTypes.IPAMSpec{
-							Pool: ipamTypes.AllocationMap{
-								"10.1.1.226": ipamTypes.AllocationIP{
-									Resource: "eni-1",
-								},
-							},
-						},
-					},
-					Status: ciliumv2.NodeStatus{
-						ENI: eniTypes.ENIStatus{
-							ENIs: map[string]eniTypes.ENI{
-								"eni-2": {
-									ID: "eni-2",
-									Addresses: []string{
-										"10.1.1.226",
-										"10.1.1.229",
-									},
-									VPC: eniTypes.AwsVPC{
-										ID:          "vpc-1",
-										PrimaryCIDR: "10.1.0.0/16",
-										CIDRs: []string{
-											"10.1.0.0/16",
-											"10.2.0.0/16",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			wantErr: true,
-			want:    "ENI eni-1 not found in status",
-		},
-		{
-			name: "ENI IP not found in status",
-			args: args{
-				node: &ciliumv2.CiliumNode{
-					Spec: ciliumv2.NodeSpec{
-						IPAM: ipamTypes.IPAMSpec{
-							Pool: ipamTypes.AllocationMap{
-								"10.1.1.227": ipamTypes.AllocationIP{
-									Resource: "eni-1",
-								},
-							},
-						},
-					},
-					Status: ciliumv2.NodeStatus{
-						ENI: eniTypes.ENIStatus{
-							ENIs: map[string]eniTypes.ENI{
-								"eni-1": {
-									ID: "eni-1",
-									Addresses: []string{
-										"10.1.1.226",
-										"10.1.1.229",
-									},
-									VPC: eniTypes.AwsVPC{
-										ID:          "vpc-1",
-										PrimaryCIDR: "10.1.0.0/16",
-										CIDRs: []string{
-											"10.1.0.0/16",
-											"10.2.0.0/16",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			wantErr: true,
-			want:    "ENI eni-1 does not have address 10.1.1.227",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := validateENIConfig(tt.args.node)
-			require.Equal(t, tt.wantErr, got != nil, "error: %v", got)
-			if tt.wantErr {
-				require.Equal(t, tt.want, got.Error())
-			}
-		})
-	}
 }
